Create APA-Style Flextable Object

Function that creates an APA-style table using flextable::flextable().

Usage

ffcar_flextable(
  dtf_table,
  chr_header_row_1,
  chr_header_row_2 = NULL,
  chr_header_row_3 = NULL,
  lst_alignment = NULL,
  int_font_size = 11,
  chr_font_family = "Arial",
  lgc_color_rows = NULL,
  chr_row_color = NULL,
  lgc_merge_header_cells = TRUE
)

Arguments

dtf_table

A data frame with the table content.

chr_header_row_1

A character vector matching in length to the number of columns of dtf_table, the headers for the columns.

chr_header_row_2

A character vector matching in length to the number of columns of dtf_table, an optional second line for the column headers.

chr_header_row_3

A character vector matching in length to the number of columns of dtf_table, an optional third line for the column headers.

lst_alignment

A named list consisting of the elements left, center, and/or right, each consisting of the column indices to apply the corresponding alignment.

int_font_size

An integer value, the font size for the table.

chr_font_family

A character string, the font family.

lgc_color_rows

A logical vector matching in length to the number of rows of dtf_table, specifying the rows to be colored.

chr_row_color

A character string giving the optional background color for rows. If no input is provided for lgc_color_rows, results in every other row being colored.

lgc_merge_header_cells

A logical value; if TRUE merges equivalent header cells into a single cell.

Value

A flextable object.

Examples

data(mtcars)
dtf_example <- mtcars[1:4, 1:3]
ft <- ffcar_flextable(
  dtf_example,
  c( 'Miles per gallon', 'Cylinders', 'Displacement' )
)
ft

Miles per gallon

Cylinders

Displacement

21

6

160

21

6

160

22.8

4

108

21.4

6

258